sms-backend/kimi_mpesa_reference/mpesa-school-saas/requirements.txt
# Django Core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.0

# Caching & Queues
redis>=4.5.0
django-redis>=5.3.0

# Celery
celery>=5.3.0
celery-beat>=2.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# HTTP Requests
requests>=2.31.0
urllib3>=2.0.0

# Environment
python-decouple>=3.8
python-dotenv>=1.0.0

# Production
 gunicorn>=21.0.0
whitenoise>=6.5.0

# API Documentation
drf-spectacular>=0.26.0

# PDF Generation (for exports)
reportlab>=4.0.0

# Monitoring & Logging
sentry-sdk>=1.30.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0

# Development
ipython>=8.14.0
django-extensions>=3.2.0

# Utilities
Pillow>=10.0.0
python-dateutil>=2.8.0
pytz>=2023.3
